Fix for problem when returning short
When returning short datatype from c# to node, happens an error 'Invalid cast', this change fix the problem.

+++ src/dotnet/clrfunc.cpp
@@ -182,7 +182,7 @@ v8::Local<v8::Value> ClrFunc::MarshalCLRToV8(System::Object^ netdata)
     }
     else if (type == System::Int16::typeid)
     {
-        jsdata = Nan::New<v8::Integer>((int)netdata);
+        jsdata = Nan::New<v8::Integer>((short)netdata);
     }
     else if (type == System::Int64::typeid)
     {
